feat: Add fast scroller to ArtistSongsScreen
This commit introduces a `VerticalFastScroller` to the `ArtistSongsScreen`.

The `LazyColumn` displaying the list of songs is now wrapped with `VerticalFastScroller`, providing users with a faster way to navigate through long lists of songs by an artist.

The fast scroller has top content padding of `16.dp` and no end content padding.